// frontend/src/types/generated/SensorReading.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SensorReadingResponse = { id: number, timestamp: string, node_id: string, soil_temperature: number, soil_moisture: number, air_temperature: number | null, air_humidity: number | null, light_level: number | null, };